skip org creation of org is empty
some IdP (e.g. FranceConnect) seem to return a blank ("") organization,
we cannot rely on the `try {} catch (NameNotFoundException)` here, but
at least we can just skip the org creation if we don't have at least a
name to create it.

Tests: tried with the IT testsuite from the branch still in review from
PR#63, but which does not exist yet ... A simple case where
"preauth-org: ''" will make it.