oauth2/oidc - avoids throwing exceptions when trying to maps roles from json payload
If the JSON path is not found into the user info's json payload, an
exception is thrown. We need to catch it and return a list of empty
role instead.

In a situation where there are several OIDC providers, we cannot ensure
that they will send the exact same json payload layout, which can lead
to errors 500 because one configuration can fit the first identity
providers but not the other ones.

Tests: make test OK, added a unit test.